Move FirstPassPrecompiledScript to precompiled groovy plugins package
